applied patches contribued by jailbird, Dustin Marquess <jailbird@alcatraz.fdf.net> for alphas 32 bit etc.

+++ src/ircsprintf.c
@@ -73,13 +73,13 @@ va_dcl
 	    rp = inp[++i];                  /* get the next parameter */
 	    break;
 	  case 'c':
-	    *wp++ = (char) ((int)rp);
+	    *wp++ = (char) ((long)rp);
 	    rp = inp[++i];
 	    break;
 	  case 'd':
 	    {
-	      register int myint;
-	      myint = (int)rp;
+	      register long myint;
+	      myint = (long)rp;
 	      
 	      if (myint < 100 || myint > 999)
 		{
@@ -102,8 +102,8 @@ va_dcl
 	  break;
 	  case 'u':
 	    {
-	      register unsigned int myuint;
-	      myuint = (unsigned int)rp;
+	      register unsigned long myuint;
+	      myuint = (unsigned long)rp;
 	  
 	      if (myuint < 100 || myuint > 999)
 		{
